--- src/assets/styles/fonts.scss
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Mulish:ital,wght@0,200..1000;1,200..1000&display=swap');
