+++ src/components/admin/admin-navigation/admin-navigation.scss
@@ -4,7 +4,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    font-size: 36px;
+    font-size: 0.625rem;
   }
 
   &-pages {
