+++ src/utils/mock-data/admin-page/teamPage.ts
@@ -1,0 +1,84 @@
+import {Member} from "../../../pages/admin/team/components/members-list/MembersList";
+
+export const mockMembers: Member[] = [
+    {
+        id: 1,
+        img: "https://randomuser.me/api/portraits/men/1.jpg",
+        fullName: "First First",
+        description: "Software EngineerS",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 2,
+        img: "https://randomuser.me/api/portraits/men/1.jpg",
+        fullName: "Second Second",
+        description: "Software EngineerSoftware Engineer",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 3,
+        img: "https://randomuser.me/api/portraits/men/1.jpg",
+        fullName: "Sohn Carter",
+        description: "Software EngineerSoftware Engineer",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 4,
+        img: "https://randomuser.me/api/portraits/women/2.jpg",
+        fullName: "Mmily Rose",
+        description: "UI/UX Designer",
+        status: "Чернетка",
+        category: "Основна команда",
+    },
+    {
+        id: 5,
+        img: "https://randomuser.me/api/portraits/men/3.jpg",
+        fullName: "Michael Blake",
+        description: "DevOps Engineer",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 6,
+        img: "https://randomuser.me/api/portraits/men/5.jpg",
+        fullName: "Daniel Li",
+        description: "QA Analyst",
+        status: "Чернетка",
+        category: "Основна команда",
+    },
+    {
+        id: 7,
+        img: "https://randomuser.me/api/portraits/women/6.jpg",
+        fullName: "Isabella Garcia",
+        description: "HR Specialist",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 8,
+        img: "https://randomuser.me/api/portraits/women/8.jpg",
+        fullName: "Olivia Brown",
+        description: "Marketing Manager",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 9,
+        img: "https://randomuser.me/api/portraits/men/9.jpg",
+        fullName: "William Smith",
+        description: "Backend Developer",
+        status: "Опубліковано",
+        category: "Основна команда",
+    },
+    {
+        id: 10,
+        img: "https://randomuser.me/api/portraits/women/10.jpg",
+        fullName: "Ava Thompson",
+        description: "Business Analyst",
+        status: "Чернетка",
+        category: "Основна команда",
+    },
+];
